hahahaahahahahahahha infinite repeating decimal silly

+++ server/app/core/types.py
@@ -134,10 +134,13 @@ class Node(BaseModel):
     prompt: str
     bbox: BoundingBox
     proxy_shape: ProxyShape | None = None
-    # Yaw, radians, world-frame rotation about +Y. Zero = the mesh's
+    # Yaw, integer degrees, world-frame rotation about +Y. Zero = the mesh's
     # canonical front (Trellis output's +Z) faces world +Z (toward viewer).
-    # Positive = right-handed yaw (front rotates toward world -X).
-    orientation: float = 0.0
+    # Positive = right-handed yaw (front rotates toward world -X). Stored
+    # as a bounded int rather than free-float radians because unconstrained
+    # float schemas let some providers loop in the exponent grammar
+    # (`0.03e-30557580…`), eating the whole token budget on a single field.
+    orientation: int = 0
     relationships: list[Relationship] = Field(default_factory=list)
     mesh_url: str | None = None
     parent_id: str | None = None
